Socket deleted. Header bar loggged-out finished.
