+++ ProjetoMinions/package.json
@@ -9,9 +9,11 @@
   "author": "EJEC Minions",
   "license": "ISC",
   "dependencies": {
-    "bootstrap": "^4.3.1",
+    "axios": "^0.18.0",
+    "body-parser": "^1.18.3",
     "express": "^4.16.4",
-    "socket.io": "^2.2.0",
+    "jshint": "^2.10.2",
+    "nodemon": "^1.18.11",
     "sqlite3": "^4.0.6"
   }
 }
