--- ProjetoMinions/sockets.js
@@ -1,10 +1,0 @@
-var socketMethods = {
-	signinValidation: function(socket, db){
-
-		let response = {user: true, cpf: true}
-
-		return response;
-	}
-}
-
-module.exports = socketMethods;
